# amd: faster xcc sync