```
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 21.72 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 21.00 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 19.44 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
```

vs

```
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 51.08 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 51.40 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
nimlgen@tinyamd2:~/tinygrad$ AMD=1 PYTHONPATH=. python3 test/test_hcq.py TestHCQ.test_speed_exec_time
exec kernel time: 50.60 us
.
----------------------------------------------------------------------
Ran 1 test in 0.031s

OK
```