@chenyuxyz 2025-04-08
tested this on bert twice and got nan loss both time fairly soon
https://wandb.ai/chenyuxyz/MLPerf-BERT/runs/ho05wonp/logs